requirements_all.txt
appdirs==1.4.4
astroid==2.4.2
certifi==2020.4.5.2
chardet==3.0.4
distlib==0.3.0
et-xmlfile==1.0.1
excel2json-3==0.1.6
filelock==3.0.12
idna==2.9
importlib-metadata==1.6.1
isort==4.3.21
jdcal==1.4.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
openpyxl==3.0.3
numpy==1.18.5
pandas==1.0.4
pylint==2.5.3
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.9
virtualenv==20.0.23
voluptuous==0.11.7
wrapt==1.12.1
xlrd==1.2.0
zipp==3.1.0
